dom/html/test/test_rowscollection.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=772869
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 772869</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=772869">Mozilla Bug 772869</a>
<p id="display"></p>
<div id="content" style="display: none">
  <table id="f">
    <thead>
      <tr id="x"></tr>
    </thead>
    <tfoot>
      <tr id="z"></tr>
      <tr id="w"></tr>
    </tfoot>
    <tr id="x"></tr>
    <tr id="y"></tr>
    <tbody>
      <tr id="z"></tr>
    </tbody>
  </table>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 772869 **/
var x = $("f").rows;
x.something = "another";
var names = [];
for (var name in x) {
  names.push(name);
}
is(names.length, 10, "Should have 10 enumerated names");
is(names[0], "0", "Enum entry 1")
is(names[1], "1", "Enum entry 2")
is(names[2], "2", "Enum entry 3")
is(names[3], "3", "Enum entry 4")
is(names[4], "4", "Enum entry 5")
is(names[5], "5", "Enum entry 6")
is(names[6], "something", "Enum entry 7")
is(names[7], "item", "Enum entry 8")
is(names[8], "namedItem", "Enum entry 9")
is(names[9], "length", "Enum entry 10");

names = Object.getOwnPropertyNames(x);
is(names.length, 11, "Should have 11 items");
is(names[0], "0", "Entry 1")
is(names[1], "1", "Entry 2")
is(names[2], "2", "Entry 3")
is(names[3], "3", "Entry 4")
is(names[4], "4", "Entry 5")
is(names[5], "5", "Entry 6")
is(names[6], "x", "Entry 7")
is(names[7], "y", "Entry 8")
is(names[8], "z", "Entry 9")
is(names[9], "w", "Entry 10")
is(names[10], "something", "Entry 11")
</script>
</pre>
</body>
</html>
